refactor: AOP 클래스명 수정, 코드 리팩토링 (#291)
