+++ server/momo-api/src/main/java/com/momo/common/aop/PerformanceAspect.java
@@ -1,4 +1,4 @@
-package com.momo.common;
+package com.momo.common.aop;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.LinkedHashMap;
@@ -20,19 +20,20 @@ import org.springframework.web.context.request.ServletRequestAttributes;
 @Order(1)
 @Component
 @RequiredArgsConstructor
-public class ElapsedTimeLogAspect {
+public class PerformanceAspect {
 
     private final ObjectMapper objectMapper;
 
-    private static final String ELAPSED_TIME_LOG_FORMAT = "Elapsed Time : {}";
+    private static final String PERFORMANCE_FORMAT = "Perform : {}";
 
     private static final double MILLI_TO_SECOND_UNIT = 0.001;
-    private static final double MAX_AFFORDABLE_TIME = 3;
+    private static final double MAX_PERFORMANCE_TIME = 3;
 
     @Around("execution(* com.momo.api..*Controller.*(..))")
     public Object printLog(ProceedingJoinPoint joinPoint) throws Throwable {
         HttpServletRequest request = ((ServletRequestAttributes) Objects.requireNonNull(
             RequestContextHolder.getRequestAttributes())).getRequest();
+
         long startTime = System.currentTimeMillis();
         Object proceed = joinPoint.proceed();
         long endTime = System.currentTimeMillis();
@@ -41,15 +42,15 @@ public class ElapsedTimeLogAspect {
         Map<String, Object> map = new LinkedHashMap<>();
         map.put("method", request.getMethod());
         map.put("uri", request.getRequestURI());
-        map.put("time", elapsedTime);
+        map.put("time", elapsedTime + "s");
 
         String resultJson = objectMapper.writerWithDefaultPrettyPrinter().writeValueAsString(map);
-        if (elapsedTime > MAX_AFFORDABLE_TIME) {
-            log.warn(ELAPSED_TIME_LOG_FORMAT, resultJson);
+        if (elapsedTime > MAX_PERFORMANCE_TIME) {
+            log.warn(PERFORMANCE_FORMAT, resultJson);
             return proceed;
         }
 
-        log.info(ELAPSED_TIME_LOG_FORMAT, resultJson);
+        log.info(PERFORMANCE_FORMAT, resultJson);
 
         return proceed;
     }
